packages/contracts/truffle/contracts/requirements.txt
appdirs==1.4.3
asn1crypto==1.4.0
attrs==20.2.0
base58==2.0.1
bitarray==1.2.2
bitcoin==1.1.42
blake2b-py==0.1.3
CacheControl==0.12.6
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
coincurve==13.0.0
colorama==0.4.3
contextlib2==0.6.0
cytoolz==0.10.1
distlib==0.3.0
distro==1.4.0
eth==0.0.1
eth-abi==2.1.1
eth-account==0.5.3
eth-bloom==1.0.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-tester==0.5.0b2
eth-typing==2.2.2
eth-utils==1.9.5
ethereum==2.3.2
future==0.18.2
hexbytes==0.2.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipfshttpclient==0.6.1
jsonschema==3.2.0
lockfile==0.12.2
lru-dict==1.1.6
msgpack==0.6.2
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==20.3
parsimonious==0.8.1
pbkdf2==1.3
pep517==0.8.2
progress==1.5
protobuf==3.13.0
py-ecc==4.0.0
py-evm==0.3.0a19
pycparser==2.20
pycryptodome==3.9.8
pyethash==0.1.27
pyparsing==2.4.6
pyrsistent==0.17.0
pysha3==1.0.2
pytoml==0.1.21
PyYAML==5.3.1
repoze.lru==0.7
requests==2.22.0
retrying==1.3.3
rlp==2.0.0a1
rusty-rlp==0.1.15
scrypt==0.8.15
semantic-version==2.8.5
six==1.14.0
sortedcontainers==2.2.2
toolz==0.10.0
trie==2.0.0a4
typing-extensions==3.7.4.3
urllib3==1.25.8
varint==1.0.2
web3==5.12.1
webencodings==0.5.1
websockets==8.1
